protect local pre-commit hooks upon githook install if they already exist for some reason

+++ git-hook/install_hook_local.sh
@@ -1,3 +1,5 @@
-mv ../.git/hooks/pre-commit ../.git/hooks/pre-commit.local 2>/dev/null
+if [ ! -f ../.git/hooks/pre-commit.local ]; then
+   mv ../.git/hooks/pre-commit ../.git/hooks/pre-commit.local 2>/dev/null
+fi
 cp pre-commit ../.git/hooks
 chmod +x ../.git/hooks/*
